Update Map and List Screen
