+++ constants/Region.js
@@ -1,0 +1,6 @@
+export const INITIAL_REGION = {
+  latitude: 42.24755809691715, 
+  longitude: -87.8291705184884,
+  latitudeDelta: 0.007,
+  longitudeDelta: 0.007,
+};
